[FEAT/#921] 셀프 리뷰 반영: 스와이프 상태 분리, 네이밍 정리, 방어 로직 보강
