+++ presentation/main/src/main/java/com/hilingual/presentation/main/MainScreen.kt
@@ -285,8 +285,10 @@ internal fun MainScreen(
 
                     vocaNavGraph(
                         paddingValues = innerPadding,
-                        navController = appState.navController,
                         navigateToHome = appState::navigateToHome,
+                        navigateToVocaReview = appState::navigateToVocaReview,
+                        navigateUp = appState::navigateUp,
+                        navigateUpWithReviewSaved = appState::navigateUpWithVocaReviewSaved,
                     )
 
                     diaryFeedbackNavGraph(
